Prometheus: VI
how many of these will there be please stop already how can one bear to endure literally 6 Prometheus redesigns, seriously there was the one that was basically a light Ares, then the one that was smaller in the weird Blue Square Vera engine era, then it became like half the size with the one that had the retractable turrets??? and after that every single one was slightly shorter. now the new one has a comically large frontal window. when will it stop???